Update Ceph-CSI manifests as per changes to ceph-csi repository
Ceph-CSI repository had a recent PR [1] merged that contains
changes to the pod manifests.

This PR to Rook updates the pod manifests to catchup with the
changes to Ceph-CSI master.

Once Ceph-CSI has a 1.1.0 release, future PRs to Rook on manifest
changes would track that branch (at least for Rook 1.1 release).

[1] PR in ceph-csi changing the manifest:
  https://github.com/ceph/ceph-csi/pull/443

Signed-off-by: ShyamsundarR <srangana@redhat.com>