enhance(ceph): Adds support for storageClassDeviceSet in ceph
Signed-off-by: Ashish Ranjan <aranjan@redhat.com>

This commit adds support for storageClassDeviceSet in rook-ceph. The primary motivation for this feature is to take advantage of the mobility of storage in cloud-based environments by defining storage based on sets of devices consumed as block-mode PVCs. In environments like AWS you can request remote storage that can be dynamically provisioned and attached to any node in a given Availability Zone (AZ).

**Note: Its currently a work in progress PR and will change according to this proposal #3306