trigger the integration tests for all branches if tags not set
The integration tests for storage providers will be skipped
for example if [test ceph] is specified. If the [test] tag
is not specified, all tests should be run. If the tag is not
specified, the Jenkinsfile will pass null, which needs to be
handled by the test code deciding to run all the tests.

Signed-off-by: travisn <tnielsen@redhat.com>