ceph: enable LoadBalancerService and LoadBalancerIP as options in Rook config
This patches https://github.com/rook/rook/issues/3404.

The gist of this PR is that the `mon` service for Ceph now allows
a) being declared as a LoadBalancer service, and b) allows operators
to supply a custom LoadBalancerIP if desired.

This effectively means cluster operators can now expose their Ceph clusters
externally from within configuration.

To do this, I've modified the `rook.NetworkSpec` interface accordingly.
This interface is currently consumed by both Cassandra and Ceph - however,
no changes to the Cassandra code have been made to honor this.

*Tests*

I noticed that we only test one particular YAML configuration of Ceph rather
than for different configuration options, and that Ceph tests are keyed only
to use the version of rook to decide which YAML to use.

This makes it prohibitive to test for multiple different YAML configurations
in the same rook version. I have accordingly left out tests for the interim
to keep changes small and reasonable.

Signed-off-by: Akshat Mahajan <akshatm.bkk@gmail.com>