mgr: set public-addr flag for MGR
This fixes that the MGR is binding to the gateway IP instead of the
actual Pod IP, as seen in #3136.
When hostNetwork: true is set the public-addrr is not set so custom
network configs can be used.

Signed-off-by: Alexander Trost <galexrt@googlemail.com>
(cherry picked from commit 6daf4087c794b90bf504bc8e7ffe04a0594f6df7)