ceph: remove default database size for OSDs
Since Rook is no longer creating partitions and leaving all that
work to ceph-volume, we should let ceph-volume choose the defaults
for the database size and use the max available. The default of 20GB
in fact doesn't even make sense since Ceph will only use the first 3GB
of that size and the other 17GB was wasted.

Signed-off-by: travisn <tnielsen@redhat.com>