ceph: null configOverride sets option to default
There was no option previously to remove a config value from Ceph's
config database via the CRD once a value had been previously set. Allow
the user to set `value` to a null value (`null` or `~` in YAML, or
`null` in JSON) to indicate that Ceph should take the default value for
an option. Also clarify that if an override is removed from the Rook
CRD, the option will retain its previously set value and that Rook will
no longer try to manage the option.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>