Increase sidecar timeout from 60s to 150s
At present, the request timeout of sidecars
are at 60s and this is a request to increase
this time out value to 150s or higher. The
higher timeout value can help to reduce the
load of our backend ceph cluster and also can
avoid throttling issues at sidecars to an extend.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>