ceph: osd init activate VGs during init
Running `vgchange -ay` will enable all VGs on the host. This will remove
the need to have the lvm* services running on the host. In such a case
with a host, this would lead to all OSDs on that host to be unavailable
and the OSD Pods crashlooping as they can't find their VG and / or LV.

Signed-off-by: Alexander Trost <galexrt@googlemail.com>