tests: by default run the minimum test matrix in PRs
The CI runs the integration tests on several versions of K8s. By default
the integration tests have all been running on all versions of K8s. With the
recent option to run a minimum matrix for ceph tests it has shown to be reliable
for PRs and is much faster to have the test distribution.

Master and release builds will still run the full set of test suites on all
versions of K8s.

Signed-off-by: travisn <tnielsen@redhat.com>